Move the intelligence part to a separate file
